requirements.txt
awsebcli==3.14.2
botocore==1.10.52
cement==2.8.2
colorama==0.3.9
Django==2.0.6
django-annoying==0.10.4
django-mysql==2.3.0
django-rest-framework==0.1.0
django-taggit==0.22.2
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
docutils==0.14
jmespath==0.9.3
Markdown==2.6.11
mysqlclient==1.3.12
pathspec==0.5.5
PyJWT==1.6.4
python-dateutil==2.7.3
pytz==2018.4
PyYAML==3.12
semantic-version==2.5.0
six==1.11.0
tabulate==0.7.5
termcolor==1.1.0
